+++ NEWS.md
@@ -2,6 +2,7 @@
 * Features: 
   - Refactor adm_lines() and obs_lines() function. adm_lines() is now based on mrgsolve::ev, and can accept "ss" specification. Covered by tests.
   - Check for mandatory columns in data set. MDV automatically supplied.  #31
+  - Check the model to see if it fills mapbayr specification
   
 * Fix bugs: 
   - Throw an error if no dataset is passed. #29
@@ -9,7 +10,6 @@
   - mbraugment with n compartments > 1 and n ID > 1 #33
 
 * Miscellaneous: 
-  - 
   - ofv computation now uses mrgsim_q() faster than basic mrgsim_df (theoretically because I did not benchmarked)
   - Added a `NEWS.md` file to track changes to the package.